refactor: Break Districts List screen into separate widgets
